requirements.txt
dbus-python==1.3.2
numpy==2.4.4
pandas==3.0.2
psycopg2-binary==2.9.11
PyGObject==3.48.2
python-dateutil==2.9.0.post0
setuptools==68.1.2
six==1.17.0
wheel==0.42.0
